renamed standard card to protocard
