+++ app/frontend/src/components/Cards/ProtoCard.tsx
