+++ app/frontend/src/components/Cards/_ProtoCard.scss
