+++ app/frontend/src/index.scss
@@ -6,8 +6,8 @@
 
 // Import component styles
 @use "components/Buttons" as *;
-@use "components/Cards/StandardCard";
 @use "components/Cards/CircleCard";
+@use "components/Cards/ProtoCard";
 @use "components/Carousel/ScrollCarousel";
 @use "components/Dialog/Dialog";
 @use "components/Inputs" as *;
